+++ src/core/geometry/BezierCurve3D.cpp
@@ -55,7 +55,7 @@ void BezierCurve3D_Geo::buildEdgeGeometries()
     }
 
     // 收集所有控制点
-    std::vector<glm::vec3> allPoints;
+    std::vector<glm::dvec3> allPoints;
     for (auto& points : controlPointss)
         for (auto& point : points)
         {
@@ -104,3 +104,6 @@ void BezierCurve3D_Geo::buildFaceGeometries()
     mm_node()->clearFaceGeometry();
     // 贝塞尔曲线没有面
 }
+
+
+
